Fixed FullScreen FX and some duplicate rendering
FullScreen FX were broken because of the make shift UI solution. It was
treating the FX as if they were UI. To fix it I had to create a separate
stereoDepthType (STEREO_DEPTH_TYPE_DISABLE), it used to be that not
having any depth is the same as being near, but this is not the case
with an HMD, so now depth must be explicitly disabled, so fullscreen fx
can do that now.

We also had an issue with the HUD and CopyRender being done twice for
both eyes because they did not have a eyeViewBuffer set. The HUD and
FullScreenFX are drawn from functions that get called for each eye, most
other UI code creates render commands only once. To fix this I added
code to specify what the viewEyeBuffer is, and for CopyRender, avoid
doing an unnecessary copy (Which helps fix some effects that try and
accumulate the screen buffer).

I also fixed the Fullscreen warp effect, by shifting it for the left and
right eyes differently.

The helltime effect is not entirely working right. It has an
accumulation buffer image, but that same buffer is used for both the
left and right images when each eye should have their own buffer. It
also may make sense to also shrink some of these effects since in VR you
have a much wider field of view than was originally intended.