Fix for BFG and refactor.
The first frame of the BFG fire animation was causing the projectile to
always fire more upward (the animation pitches the weapon to represent
kickback). To fix this I consolidated the muzzle getting code (also used
for laserpointer), and just use the muzzle from the frame before the
animation, that does not yet have the kickback. This really seems like
an animation bug (either in the asset or in the pipeline or maybe in the
animation system), the first frame maybe should not have the kickback in
it already, but in either case the code fix works and maybe is the best
way to do it since it gets rid of the redudant calls of getting the
muzzle position but at the same time creates a depedency risk (the
muzzle should be updated first before using it).