Added a skin wrapper for VR.
This is a little hacky. Basically a skin that removes the arms from the
weapons, but wraps other skins (like the invisible skins).

Would be less hacky if I acquired the materials early on during level
loading (it throws an exception if you try to load a material from the
game thread), and would be less hacky if it wasn't done as global but as
a member variable of idWeapon.