Control Changes
Swapped jump and reload buttons, it felt more natural to reload with
your gun hand (right hand) and to jump with you navigation hand (left
hand). Might try swapping the jump and flashlight button, if it seems
kind of tough to jump that way on various controllers.

Navigating now moves in the direction of the facing of your left hand
controller.