Input Refactor
I think this is the more ideal way input should be done in this engine,
through user commands. This also may allow for multiplayer in the
future, but latency will be a huge deal, there could also be latency
issues with doing it this way since the game thread runs 50% faster than
the render thread.

I needed a conceptual direction of the body, so I chose the left
controller, it makes sense since it also controls the bodies locomotion.
This allows players to move in the direction they are facing, both
physically and with the controller.

More tweaks and bug may need to be done.