Fix for Cinematics 3
Hopefully the final fix. The same issue with the left controller causing
weird headmovements can occur when the pitch is not zero, so we remove
the pitch before mixing it with the cinematic camera. Also decided to
pull the changes out of the camera function and simply take the camera
functions output and apply against our axis matrix.