Self Shadow Experiment
This changeset adds support for self shadowing (for materials where it
is disabled), and includes a shader adjustment to reduce the popping
artifacts caused when the real normal is facing away but the smoothed
normal is facing towards the light.

It runs really fast so it's enabled by default, can be disabled in the
console with r_selfShadow or by enabling r_useShadowMapping.

The shader adjustment can be disabled in console by setting
r_selfShadowAdjust to 0.

I may disable it by default in the future after some testing, or I may
try a different approach to fixing this issue.

I am hijacking USE_GPU_SKINNING in order for the shader adjustment to
only be applied to characters, it unnecessary makes everything else look
darker when we only need to do this for smoothed objects.