Cinematic Initial View Angle Fix
Sometimes on the start of a cinematic the view angle will be incorrect,
this is because body facing ("viewAngles") is based upon the left
controller direction. To work around this, I add the hmdAxis yaw to the
lastViewAngles, so when the cinematic starts it will include the facing
direction of the hmd.