Mark different kinds of 2D graphics, and entering/exiting the shell.
This change marks different kinds of 2D graphics.

GUIMODE_SHELL  for Main screen, Pause screen, PDA/object screen.
GUIMODE_HUD for in game HUD.
GUIMODE_FULLSCREEN for Special FX, and the screen fade.

And anytime we enter a shell we capture the vr headset.

I have code that uses this information but it is still in development.