Initial cinematic positioning should just use yaw.
It's a bit weird to be looking up when a cinematic starts and for the
whole sequence to occur above you. Plus I want to emulate what the shell
does when it comes up, and it also ignores your heads pitch and roll, it
may be important to an experiment I am considering in the future. Even
the shell is intended to emulate what SteamVR dashboard does, which also
only uses your heads yaw.