GUI control updates, shell mouse head control.
Use your heads facing direction to control the mouse cursor! While more
like Oculus Home than SteamVR dashboard, this is consistent with the way
you interact with computer screens in-game. The right trigger is the
mouse button. The left trigger is the 'B' button (backs out of a menus).

Switched dialogs to behave like the shell instead of the hud, so they
can work with the mouse cursor.

After playing with the buttons in the GUI for sometime, I decided that
the right stick button (right touch button) will be the 'A' button
(selects a menu item) and the left stick button (left touch button) will
be the 'B' button (backs out of a menu).

Fixed a bug with dialogs where pressing right would cause it to keep
moving right instead of just moving right once. Needed to only send the
event once when the state changed instead of sending it each frame.