SSAO and MotionBlur Fix.
Needed to change depth buffer type to include stencil.

The way to determine if a viewDef is 3d or 2d has changed, so updated
many functions.

As a result SSAO and Motion Blur work again but are NOT RECOMMENDED to
use at all.

For SSAO there isn't enough ambient to make a difference. It destroys
game mechanics and the mood of the game to boost ambient just to show it
off. (although it looks cool in debug)

And for Motion Blur it doesn't look good with head tracking because it
doesn't take into account that your eyes don't always turn with your
head and are likely tracking things as your head moves, so it looks
funny to see things blur when turning your head. It does work well when
you use the game controller or mouse to turn.  Maybe with foviated eye
tracking, motion blur could work with head tracking.