Resolution Scaling
Added console command vr_resolutionScale, but it doesn't work realtime,
see: https://github.com/ValveSoftware/openvr/issues/216

Could try to allocate a larger buffer, and see if I can simply use less
of it, depending on scale. My understand is this is the preferred way of
doing this because it can allow a game to adapt to drops in performance
by lowering the resolution, but this is problematic in the current
architecture of this engine, since it is designed to reallocate a bunch
of buffers when the resolution changes.

Anyway, for now you have to restart after changing this.