Fixed a height and eye view issue.
There is a head tilt that I was removing but in removing it I was not
re-adding the height difference, so instead I just prevent the tilt to
begin with.

There was an additional offset in the eye matrix, this wasn't present
for the Rift, but is in the Vives matrix. This likely caused extra head
movement when turning your head (the distance was 1.5 cm). Fixing this
could possibly reduce motion sickness for some.