DPad style ui control.
Using touchpad to control the ui was pretty bad, this now requires you
to press it in like a DPAD.

Not sure if I want to make an additional change, I keep wanting to click
the middle as if that selects. Right now selecting is pressing in the
right touchpad.