Disable Grabber Screen Warp
The grabber has two warps, one on the beam and another that warps the
middle of the screen. The screen warp was kind of okay in nonVR because
the gun is always facing the same direction, but in VR when you can move
your hands independently it looks weird that it still warps the screen.