Window Output Modes
This allows for a better look at graphics for left and right images for
debugging graphics issues or streaming.

The cover mode, will only update the window a few times and stop, this
may be useful for improving performance that scale the resolution
higher. This also might help those who may come across vsync issues.