Move Mode Updates
vr_moveSpeed affects crouching speeds less, and applies to new move
modes 10 to 16.

Added new move modes 10 to 16 which are the same as 0 to 6 but are
scaled by vr_moveSpeed.

Added modes 7 8 and 9, which have 3 move speeds. A move speed scaled
down by vr_moveSpeed and then the default walk speed, and then the run
speed.